AudioProcessorParameter: Move impl to dedicated file
