+++ examples/DemoRunner/Builds/Android/app/CMakeLists.txt
@@ -812,6 +812,7 @@ add_library( ${BINARY_NAME}
     "../../../../../modules/juce_audio_processors/processors/juce_AudioProcessorGraph.cpp"
     "../../../../../modules/juce_audio_processors/processors/juce_AudioProcessorGraph.h"
     "../../../../../modules/juce_audio_processors/processors/juce_AudioProcessorListener.h"
+    "../../../../../modules/juce_audio_processors/processors/juce_AudioProcessorParameter.cpp"
     "../../../../../modules/juce_audio_processors/processors/juce_AudioProcessorParameter.h"
     "../../../../../modules/juce_audio_processors/processors/juce_AudioProcessorParameterGroup.cpp"
     "../../../../../modules/juce_audio_processors/processors/juce_AudioProcessorParameterGroup.h"
@@ -3474,6 +3475,7 @@ set_source_files_properties(
     "../../../../../modules/juce_audio_processors/processors/juce_AudioProcessorGraph.cpp"
     "../../../../../modules/juce_audio_processors/processors/juce_AudioProcessorGraph.h"
     "../../../../../modules/juce_audio_processors/processors/juce_AudioProcessorListener.h"
+    "../../../../../modules/juce_audio_processors/processors/juce_AudioProcessorParameter.cpp"
     "../../../../../modules/juce_audio_processors/processors/juce_AudioProcessorParameter.h"
     "../../../../../modules/juce_audio_processors/processors/juce_AudioProcessorParameterGroup.cpp"
     "../../../../../modules/juce_audio_processors/processors/juce_AudioProcessorParameterGroup.h"
